Add files via upload
K-fold cross-validation is a technique used to evaluate the performance of a machine learning model. The basic idea is to split your data into K equally sized "folds." The model is trained on K-1 folds and tested on the remaining fold. This process is repeated K times, each time with a different fold as the test set. The results are then averaged to give a more robust estimate of model performance.