Add files via upload
A Decision Tree is a popular machine learning algorithm used for both classification and regression tasks. It is a flowchart-like structure where internal nodes represent decisions or tests on attributes, branches represent outcomes of these decisions, and leaf nodes represent the final outcome or prediction. Decision Trees are intuitive and widely used due to their interpretability.